+++ src/components/QRBarCode/QRBarCode.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import {
   View,
@@ -7,18 +7,18 @@ import {
   TouchableOpacity,
   ScrollView,
   Platform,
-  Alert
+  Alert,
 } from 'react-native';
 import ScreenBrightness from 'react-native-screen-brightness';
 import {
   check,
   PERMISSIONS,
   RESULTS,
-  openSettings
+  openSettings,
 } from 'react-native-permissions';
 import QRCodeScanner from 'react-native-qrcode-scanner';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
-import { Actions, ActionConst } from 'react-native-router-flux';
+import {Actions, ActionConst} from 'react-native-router-flux';
 import QRCode from 'react-native-qrcode-svg';
 import Barcode from 'react-native-barcode-builder';
 import appConfig from 'app-config';
@@ -26,6 +26,8 @@ import timer from 'react-native-timer';
 import Button from 'react-native-button';
 import store from 'app-store';
 import EventTracker from '../../helper/EventTracker';
+import {APIRequest} from 'src/network/Entity';
+import APIHandler from 'src/network/APIHandler';
 
 const MAXIMUM_LUMINOUS = 0.7;
 const MIN_LUMINOUS = 0.5;
@@ -34,7 +36,7 @@ const isIos = Platform.OS === 'ios';
 
 class QRBarCode extends Component {
   static propTypes = {
-    mobxStore: PropTypes.object
+    mobxStore: PropTypes.object,
   };
 
   constructor(props) {
@@ -48,14 +50,18 @@ class QRBarCode extends Component {
       title: props.title || props.t('common:screen.qrBarCode.mainTitle'),
       content: props.content ? props.content : props.t('content.description'),
       originLuminous: MIN_LUMINOUS,
-      permissionCameraGranted: undefined
+      permissionCameraGranted: undefined,
     };
 
     this.unmounted = false;
     this.eventTracker = new EventTracker();
+    this.checkProductRequest = new APIRequest();
+    this.requests = [this.checkProductRequest];
   }
 
   componentDidMount() {
+    this.checkProductCode('abc');
+
     if (!this.props.address) {
       this._getData();
       this.setTimmer();
@@ -74,6 +80,7 @@ class QRBarCode extends Component {
     timer.clearTimeout(this, 'barcodeupdate');
     ScreenBrightness.setBrightness(this.state.originLuminous);
     this.eventTracker.clearTracking();
+    cancelRequests(this.requests);
   }
 
   async checkPermissions() {
@@ -82,12 +89,12 @@ class QRBarCode extends Component {
       permissionCameraGranted !== this.state.permissionCameraGranted &&
       !this.unmounted
     ) {
-      this.setState({ permissionCameraGranted });
+      this.setState({permissionCameraGranted});
     }
   }
 
   checkCameraPermission = async () => {
-    const { t } = this.props;
+    const {t} = this.props;
     if (!isAndroid && !isIos) {
       Alert.alert(t('common:system.camera.error.notSupport'));
       return false;
@@ -103,12 +110,12 @@ class QRBarCode extends Component {
         case RESULTS.UNAVAILABLE:
           Alert.alert(t('common:system.camera.error.unavailable'));
           console.log(
-            'This feature is not available (on this device / in this context)'
+            'This feature is not available (on this device / in this context)',
           );
           return false;
         case RESULTS.DENIED:
           console.log(
-            'The permission has not been requested / is denied but requestable'
+            'The permission has not been requested / is denied but requestable',
           );
           return false;
         case RESULTS.GRANTED:
@@ -126,10 +133,10 @@ class QRBarCode extends Component {
   };
 
   handleBrightness = () => {
-    ScreenBrightness.getBrightness().then(originLuminous => {
+    ScreenBrightness.getBrightness().then((originLuminous) => {
       if (originLuminous < MIN_LUMINOUS) {
-        this.setState({ originLuminous }, () =>
-          ScreenBrightness.setBrightness(MAXIMUM_LUMINOUS)
+        this.setState({originLuminous}, () =>
+          ScreenBrightness.setBrightness(MAXIMUM_LUMINOUS),
         );
       }
     });
@@ -143,17 +150,17 @@ class QRBarCode extends Component {
         this._getData();
         this.setTimmer();
       },
-      5000
+      5000,
     );
   }
 
   async _getData() {
-    this.setState({ loading: true });
+    this.setState({loading: true});
     const response = await APIHandler.user_barcode(
-      this.props.mobxStore.store_id
+      this.props.mobxStore.store_id,
     );
     if (response && response.status == STATUS_SUCCESS && !this.unmounted) {
-      this.setState({ barcode: response.data.barcode });
+      this.setState({barcode: response.data.barcode});
     }
   }
 
@@ -170,7 +177,7 @@ class QRBarCode extends Component {
     }
     this.setState(
       {
-        loading: true
+        loading: true,
       },
       async () => {
         try {
@@ -180,25 +187,25 @@ class QRBarCode extends Component {
               action(() => {
                 this.setState(
                   {
-                    loading: false
+                    loading: false,
                   },
                   () => {
                     setTimeout(() => {
                       !this.unmounted &&
                         this._proccessQRCodeResult(response.data.barcode);
                     }, 450);
-                  }
+                  },
                 );
               })();
             } else {
               action(() => {
                 this.setState(
                   {
-                    loading: false
+                    loading: false,
                   },
                   () => {
                     this._open_webview(link);
-                  }
+                  },
                 );
               })();
             }
@@ -208,20 +215,20 @@ class QRBarCode extends Component {
             !this.unmounted &&
               this.setState(
                 {
-                  loading: false
+                  loading: false,
                 },
                 () => {
                   this._open_webview(link);
-                }
+                },
               );
           })();
         } finally {
           !this.unmounted &&
             this.setState({
-              loading: false
+              loading: false,
             });
         }
-      }
+      },
     );
   }
 
@@ -229,11 +236,11 @@ class QRBarCode extends Component {
    * Lay tai khoan tu Ma Tai khoan
    */
   _getAccountByBarcode(barcode) {
-    const { t } = this.props;
+    const {t} = this.props;
     //, password, refer
     this.setState(
       {
-        loading: true
+        loading: true,
       },
       async () => {
         try {
@@ -243,7 +250,7 @@ class QRBarCode extends Component {
               action(() => {
                 this.setState(
                   {
-                    loading: false
+                    loading: false,
                   },
                   () => {
                     Actions.pay_account({
@@ -252,14 +259,14 @@ class QRBarCode extends Component {
                       wallet: response.data.account.default_wallet,
                       account: response.data.account,
                       app: response.data.app,
-                      type: ActionConst.REPLACE
+                      type: ActionConst.REPLACE,
                     });
 
                     flashShowMessage({
                       type: 'success',
-                      message: response.message
+                      message: response.message,
                     });
-                  }
+                  },
                 );
               })();
             } else {
@@ -270,10 +277,10 @@ class QRBarCode extends Component {
         } finally {
           !this.unmounted &&
             this.setState({
-              loading: false
+              loading: false,
             });
         }
-      }
+      },
     );
   }
 
@@ -284,12 +291,12 @@ class QRBarCode extends Component {
     //,
     this.setState(
       {
-        loading: true
+        loading: true,
       },
       async () => {
         try {
           // const data = isWalletAddressWithZoneCode(barcode);
-          const data = { qrcode: barcode };
+          const data = {qrcode: barcode};
           const response = await APIHandler.user_process_qrcode(data);
           console.log(response);
           if (
@@ -299,11 +306,11 @@ class QRBarCode extends Component {
           ) {
             this.setState(
               {
-                loading: false
+                loading: false,
               },
               () => {
-                const { t } = this.props;
-                const { wallet, to_wallet, site: receiverInfo } = response.data;
+                const {t} = this.props;
+                const {wallet, to_wallet, site: receiverInfo} = response.data;
                 let receiverTel = receiverInfo.tel;
                 receiverTel = receiverTel.split('+84').join('0');
                 if (receiverTel.slice(0, 2) === '84') {
@@ -313,7 +320,7 @@ class QRBarCode extends Component {
                 console.log(response.data);
                 Actions.push(appConfig.routes.transferPayment, {
                   title: t('transfer:transferPaymentTitle', {
-                    walletName: receiverInfo.name
+                    walletName: receiverInfo.name,
                   }),
                   wallet,
                   showWallet: true,
@@ -326,8 +333,8 @@ class QRBarCode extends Component {
                     originTel: receiverInfo.tel,
                     avatar: receiverInfo.logo_url,
                     address: receiverInfo.address,
-                    notInContact: true
-                  }
+                    notInContact: true,
+                  },
                 });
                 // Actions.push(appConfig.routes.payWallet, {
                 //   title: "Chuyển khoản",
@@ -336,7 +343,7 @@ class QRBarCode extends Component {
                 //   type: ActionConst.REPLACE,
                 // });
                 Toast.show(response.message, Toast.SHORT);
-              }
+              },
             );
           } else {
             this._search_store(barcode);
@@ -344,7 +351,7 @@ class QRBarCode extends Component {
         } catch (e) {
           this._search_store(barcode);
         }
-      }
+      },
     );
   }
 
@@ -357,21 +364,21 @@ class QRBarCode extends Component {
     // hide tutorial go store
     if (this.props.that) {
       this.props.that.setState({
-        show_go_store: false
+        show_go_store: false,
       });
     }
 
     Actions.push(appConfig.routes.store, {
       title: item.name,
       goCategory: category_id,
-      type: ActionConst.REPLACE
+      type: ActionConst.REPLACE,
     });
   }
   _goProduct(item) {
     Actions.item({
       title: item.name,
       item,
-      type: ActionConst.REPLACE
+      type: ActionConst.REPLACE,
     });
   }
 
@@ -379,7 +386,7 @@ class QRBarCode extends Component {
     Actions.notify_item({
       title: item.title,
       data: item,
-      type: ActionConst.REPLACE
+      type: ActionConst.REPLACE,
     });
   }
 
@@ -390,7 +397,7 @@ class QRBarCode extends Component {
     //,
     this.setState(
       {
-        loading: true
+        loading: true,
       },
       async () => {
         try {
@@ -401,7 +408,7 @@ class QRBarCode extends Component {
               action(() => {
                 this.setState(
                   {
-                    loading: false
+                    loading: false,
                   },
                   () => {
                     if (response.data.object.type == OBJECT_TYPE_KEY_USER) {
@@ -436,7 +443,7 @@ class QRBarCode extends Component {
                       setTimeout(() => {
                         this._goStores(
                           response.data.item,
-                          response.data.item.site_product_category_id
+                          response.data.item.site_product_category_id,
                         );
                       }, 0);
                     } else if (
@@ -460,7 +467,7 @@ class QRBarCode extends Component {
                       setTimeout(() => {
                         Actions.view_orders_item({
                           data: response.data.item,
-                          title: '#' + response.data.item.cart_code
+                          title: '#' + response.data.item.cart_code,
                         });
                       }, 0);
                     } else if (
@@ -470,7 +477,7 @@ class QRBarCode extends Component {
                       setTimeout(() => {
                         Actions.push(appConfig.routes.voucherDetail, {
                           title: response.data.item.title,
-                          campaignId: response.data.item.id
+                          campaignId: response.data.item.id,
                         });
                       }, 0);
                     } else {
@@ -480,7 +487,7 @@ class QRBarCode extends Component {
                       }, 0);
                     }
                     // Toast.show(response.message, Toast.SHORT);
-                  }
+                  },
                 );
               })();
             } else {
@@ -490,7 +497,7 @@ class QRBarCode extends Component {
         } catch (e) {
           this._search_store(barcode);
         }
-      }
+      },
     );
   }
 
@@ -501,7 +508,7 @@ class QRBarCode extends Component {
     //,
     this.setState(
       {
-        loading: true
+        loading: true,
       },
       async () => {
         try {
@@ -510,17 +517,17 @@ class QRBarCode extends Component {
             action(() => {
               this.setState(
                 {
-                  loading: false
+                  loading: false,
                 },
                 () => {
                   Actions.view_orders_item({
                     data: response.data,
                     title: '#' + barcode,
                     tel: response.data.tel,
-                    type: ActionConst.REPLACE
+                    type: ActionConst.REPLACE,
                   });
                   // Toast.show(response.message, Toast.SHORT);
-                }
+                },
               );
             })();
           } else {
@@ -529,12 +536,12 @@ class QRBarCode extends Component {
         } catch (e) {
           this._search_store(barcode);
         }
-      }
+      },
     );
   }
 
   _search_store(barcode) {
-    const { t } = this.props;
+    const {t} = this.props;
     alert(t('invalidQRCode'));
   }
 
@@ -543,13 +550,13 @@ class QRBarCode extends Component {
     setTimeout(() => {
       Actions.webview({
         title: link,
-        url: link
+        url: link,
       });
     }, 0);
   }
 
   _proccessQRCodeResult(text_result) {
-    const { wallet, from } = this.state;
+    const {wallet, from} = this.state;
     if (text_result) {
       if (isURL(text_result)) {
         if (isLinkTickID(text_result)) {
@@ -579,26 +586,63 @@ class QRBarCode extends Component {
       } else if (isWalletAddressWithZoneCode(text_result)) {
         this._getWalletByAddressAndZoneCode(text_result);
       } else {
-        this._search_store(text_result);
+        // this._search_store(text_result);
+        this.checkProductCode(text_result);
       }
     }
   }
 
+  async checkProductCode(qrcode) {
+    const data = {qrcode};
+    const {t} = this.props;
+
+    try {
+      this.checkProductRequest.data = APIHandler.user_check_product_code(data);
+      const response = await this.checkProductRequest.promise();
+      console.log(response);
+      
+      if (response) {
+        if (response.status === STATUS_SUCCESS && response.data) {
+          Actions.push(appConfig.routes.item, {
+            item: response.data,
+            title: response.data.name,
+          });
+        } else {
+          Actions.pop();
+          Actions.push(appConfig.routes.searchStore, {
+            qr_code: qrcode,
+          });
+        }
+      } else {
+        flashShowMessage({
+          type: 'danger',
+          message: t('common:api.error.message'),
+        });
+      }
+    } catch (error) {
+      console.log('check_product_code', error);
+      flashShowMessage({
+        type: 'danger',
+        message: t('common:api.error.message'),
+      });
+    }
+  }
+
   goToSetting() {
-    const { t } = this.props;
+    const {t} = this.props;
     openSettings().catch(() =>
-      Alert.alert(t('common:system.settings.error.accessProblem'))
+      Alert.alert(t('common:system.settings.error.accessProblem')),
     );
   }
 
-  goToPayment = wallet_address => {
-    const { t } = this.props;
-    this.getUserInfo(wallet_address, receiverInfo => {
+  goToPayment = (wallet_address) => {
+    const {t} = this.props;
+    this.getUserInfo(wallet_address, (receiverInfo) => {
       Actions.pop();
       setTimeout(() => {
         const wallet = this.props.wallet || store.user_info.default_wallet;
         Actions.push(appConfig.routes.transferPayment, {
-          title: t('transferPaymentTitle', { walletName: wallet.name }),
+          title: t('transferPaymentTitle', {walletName: wallet.name}),
           wallet,
           receiver: {
             id: receiverInfo.id,
@@ -608,16 +652,16 @@ class QRBarCode extends Component {
             tel: receiverInfo.tel,
             originTel: receiverInfo.tel,
             avatar: receiverInfo.avatar,
-            notInContact: true
-          }
+            notInContact: true,
+          },
         });
       });
     });
   };
 
   getUserInfo = async (wallet_address, callBackSuccess) => {
-    const data = { wallet_address };
-    this.setState({ loading: true });
+    const data = {wallet_address};
+    this.setState({loading: true});
 
     try {
       const response = await APIHandler.user_get_info_by_wallet_address(data);
@@ -628,7 +672,7 @@ class QRBarCode extends Component {
             name: response.data.name,
             wallet_address: response.data.wallet_address,
             avatar: response.data.img,
-            tel: response.data.tel
+            tel: response.data.tel,
           });
         } else {
           Alert.alert(response.message);
@@ -637,17 +681,17 @@ class QRBarCode extends Component {
     } catch (error) {
       console.log(error);
     } finally {
-      !this.unmounted && this.setState({ loading: false });
+      !this.unmounted && this.setState({loading: false});
     }
   };
 
   renderQRCodeScanner(text_result) {
-    const { t } = this.props;
+    const {t} = this.props;
     return (
       <QRCodeScanner
         checkAndroid6Permissions={true}
-        ref={node => (this.scanner = node)}
-        onRead={e => {
+        ref={(node) => (this.scanner = node)}
+        onRead={(e) => {
           this._proccessQRCodeResult(e.data);
         }}
         topContent={
@@ -660,8 +704,7 @@ class QRBarCode extends Component {
               <Button
                 containerStyle={styles.permissionNotGrantedBtn}
                 style={styles.permissionNotGrantedSetting}
-                onPress={this.goToSetting.bind(this)}
-              >
+                onPress={this.goToSetting.bind(this)}>
                 {t('settingLabel')}
               </Button>
             </View>
@@ -679,12 +722,12 @@ class QRBarCode extends Component {
   }
 
   renderMyQRCode() {
-    const { barcode } = this.state;
-    const { t } = this.props;
+    const {barcode} = this.state;
+    const {t} = this.props;
     return (
-      <ScrollView style={{ flex: 1 }}>
+      <ScrollView style={{flex: 1}}>
         <Text style={styles.headerText}>{' ' + t('POSGuideMessage')}</Text>
-        <View style={{ marginLeft: 30, marginRight: 30 }}>
+        <View style={{marginLeft: 30, marginRight: 30}}>
           <Barcode
             value={barcode}
             format="CODE128"
@@ -696,13 +739,13 @@ class QRBarCode extends Component {
         <Text style={styles.barcodeText}>{barcode}</Text>
         <View style={styles.qrCodeView}>
           <QRCode
-            style={{ flex: 1 }}
+            style={{flex: 1}}
             value={barcode}
             size={appConfig.device.width / 3}
             logoBackgroundColor="transparent"
           />
         </View>
-        <Text style={[styles.barcodeText, { fontSize: 16 }]}>
+        <Text style={[styles.barcodeText, {fontSize: 16}]}>
           <Icon name="reload" size={16} color="#000" />
           {t('reload')}
         </Text>
@@ -713,13 +756,13 @@ class QRBarCode extends Component {
   }
 
   renderOnlyQRCode() {
-    const { barcode, content } = this.state;
+    const {barcode, content} = this.state;
     return (
-      <ScrollView style={{ flex: 1 }}>
+      <ScrollView style={{flex: 1}}>
         <Text style={styles.addressText}>{barcode}</Text>
         <View style={styles.addressQrCodeView}>
           <QRCode
-            style={{ flex: 1 }}
+            style={{flex: 1}}
             value={barcode}
             size={appConfig.device.width / 1.5}
             logoBackgroundColor="transparent"
@@ -732,16 +775,16 @@ class QRBarCode extends Component {
 
   onPressTabButton(index) {
     if (index == 0) {
-      Actions.refresh({ title: this.state.title, address: this.props.address });
+      Actions.refresh({title: this.state.title, address: this.props.address});
     } else {
-      const { t } = this.props;
-      Actions.refresh({ title: t('common:screen.qrBarCode.scanTitle') });
+      const {t} = this.props;
+      Actions.refresh({title: t('common:screen.qrBarCode.scanTitle')});
     }
-    this.setState({ index: index });
+    this.setState({index: index});
   }
 
   render() {
-    const { index, title } = this.state;
+    const {index, title} = this.state;
     return (
       <View style={styles.container}>
         <View style={styles.contentView}>
@@ -755,8 +798,7 @@ class QRBarCode extends Component {
           <TouchableOpacity
             style={styles.bottomButton}
             onPress={() => this.onPressTabButton(0)}
-            activeOpacity={1}
-          >
+            activeOpacity={1}>
             <Icon
               name="barcode-scan"
               size={20}
@@ -765,17 +807,15 @@ class QRBarCode extends Component {
             <Text
               style={[
                 styles.titleBottomButton,
-                index == 0 ? { color: global.DEFAULT_COLOR } : { color: '#000' }
-              ]}
-            >
+                index == 0 ? {color: global.DEFAULT_COLOR} : {color: '#000'},
+              ]}>
               {title}
             </Text>
           </TouchableOpacity>
           <TouchableOpacity
             style={styles.bottomButton}
             onPress={() => this.onPressTabButton(1)}
-            activeOpacity={1}
-          >
+            activeOpacity={1}>
             <Icon
               name="qrcode-scan"
               size={20}
@@ -784,9 +824,8 @@ class QRBarCode extends Component {
             <Text
               style={[
                 styles.titleBottomButton,
-                index == 1 ? { color: global.DEFAULT_COLOR } : { color: '#000' }
-              ]}
-            >
+                index == 1 ? {color: global.DEFAULT_COLOR} : {color: '#000'},
+              ]}>
               Scan QRCode
             </Text>
           </TouchableOpacity>
@@ -801,22 +840,22 @@ const styles = StyleSheet.create({
     flex: 1,
     marginBottom: 0,
     width: '100%',
-    backgroundColor: '#fff'
+    backgroundColor: '#fff',
   },
   topContent: {
     width: appConfig.device.width,
     paddingVertical: 16,
-    alignItems: 'center'
+    alignItems: 'center',
   },
   centerText: {
     lineHeight: 20,
     fontSize: 16,
     color: '#404040',
     marginLeft: 8,
-    paddingHorizontal: 15
+    paddingHorizontal: 15,
   },
   contentView: {
-    height: Util.size.height - 49 - global.NAV_HEIGHT
+    height: Util.size.height - 49 - global.NAV_HEIGHT,
   },
   bottomView: {
     backgroundColor: '#fff',
@@ -826,32 +865,32 @@ const styles = StyleSheet.create({
     right: 0,
     flexDirection: 'row',
     paddingVertical: 15,
-    borderTopWidth: .5,
-    borderColor: '#eee'
+    borderTopWidth: 0.5,
+    borderColor: '#eee',
   },
   bottomButton: {
     flex: 1,
     alignItems: 'center',
     alignContent: 'center',
-    justifyContent: 'center'
+    justifyContent: 'center',
   },
   titleBottomButton: {
     color: '#000',
     fontSize: 14,
-    marginTop: 2
+    marginTop: 2,
   },
   headerText: {
     fontSize: 15,
     color: '#000',
     marginTop: 20,
     marginLeft: 30,
-    marginRight: 20
+    marginRight: 20,
   },
   qrCodeView: {
     marginTop: 0,
     width: appConfig.device.width / 3,
     height: appConfig.device.width / 3,
-    alignSelf: 'center'
+    alignSelf: 'center',
   },
   barcodeText: {
     fontSize: 20,
@@ -861,14 +900,14 @@ const styles = StyleSheet.create({
     marginTop: 20,
     marginBottom: 20,
     marginLeft: 20,
-    marginRight: 20
+    marginRight: 20,
   },
   descText: {
     fontSize: 14,
     color: '#000',
     marginTop: 20,
     marginLeft: 70,
-    marginRight: 70
+    marginRight: 70,
   },
   addressText: {
     fontSize: 16,
@@ -878,32 +917,32 @@ const styles = StyleSheet.create({
     marginLeft: 20,
     marginRight: 20,
     textAlign: 'center',
-    fontWeight: 'bold'
+    fontWeight: 'bold',
   },
   poweredText: {
     fontSize: 14,
     color: '#000',
     bottom: 20,
-    textAlign: 'center'
+    textAlign: 'center',
   },
   addressQrCodeView: {
     marginTop: 0,
     marginBottom: 20,
     width: appConfig.device.width / 1.5,
     height: appConfig.device.width / 1.5,
-    alignSelf: 'center'
+    alignSelf: 'center',
   },
   permissionNotGrantedBtn: {
     marginTop: 15,
     backgroundColor: '#d9d9d9',
     paddingHorizontal: 10,
     paddingVertical: 5,
-    borderRadius: 5
+    borderRadius: 5,
   },
   permissionNotGrantedSetting: {
     fontSize: 14,
-    color: '#404040'
-  }
+    color: '#404040',
+  },
 });
 
 export default withTranslation(['qrBarCode', 'transfer', 'common'])(QRBarCode);