+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [r11.3.10] - 2021-02-26
+### Added
+- Logic for config key allow site sale out inventory.
+- Check product code by qrcode.
+### Fixed
+- Cart footer UI quantity.
+
 ## [r11.3.9] - 2021-02-23
 ### Add
 - Warehouse btn in detail product.